// src/enums/task.ts
export enum TaskCategory {
  DESIGN = "Design",
  DEVELOPMENT = "Utvikling",
  ECONOMY = "Økonomi",
  GAMES = "Spill",
  MARKETING = "Markedsføring"
}

export enum TaskPriority {
  HIGH = "Høy",
  MEDIUM = "Medium",
  LOW = "Lav"
}

export enum TaskStatus {
  PENDING = "Ikke startet",
  DOING = "Pågår",
  DONE = "Ferdig"
}